Added xls and xlsx import tests
